[MM-15589] Create Issue: (Jira server) Priority field fails to select (#81)
* The priority field type, inside the fields object, is an object, not a string value.
For the priority ReactSelectSetting component, filter against
this.props.value.id to get the value, not this.props.value which is an
object

* Remove default values in initialState object.
Remove special casing for priority field in JiraField component. Note
that fields that have field.allowedValues will return an object.
Otherwise returns a string with is handled when field.schema.type ===
'string'

* minor changes

* fixup