Properly load Atlassian security context in LoadInstance
There was a typo before `json:"none"` rather than `json:"-"`, leading to
the context getting (re-)stored in the DB unintentionally.

The intention was to store the original JSON regardless of what fields
we extract from it in run-time, now properly implemented.

Also updated the commented debug code in command.go so it compiles.