Try to parse events without issue_event_type_name
Some versions of Jira don't include a issue_event_type_name field in
their webhook payloads. For these events, we check for the presence of
the changelog or comment fields and handle them appropriately. This
won't help for jira:issue_updated events sent in response to a comment
deletion, but those have no real indication as to what happened anyway.